made images folder in Spinner, moved it to common
