+++ src/common/Spinner/index.js
@@ -1,0 +1,3 @@
+import { StyledSpinner } from "./styled";
+
+export const Spinner = () => <StyledSpinner />;
