--- src/feautures/Search/Spinner/index.js
@@ -1,7 +1,0 @@
-import { StyledSpinner } from "./styled";
-
-export const Spinner = () => (
-    <>
-        <StyledSpinner />
-    </>
-);
